[IMP] purchase_security_advanced: purchase representative full support
With this major enhacement, a Purchase Representative for a partner is
added. When a partner is selected for a Purchase Order, PO purchase
representative obtained from partner is automatically filled.

Also, this new field is taken in account preserving access to partner
within POs for those users that have less access to Purchases.